% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allPart.R
\name{allPart}
\alias{allPart}
\title{Concatenate identical properties of different Swift/Gadget particle species}
\usage{
allPart(dat, field = "Coordinates", species = seq(0, 5))
}
\arguments{
\item{dat}{a list containing Swift/Gadget particle data. It must have at least one sublist PartType# (with #=0,1,...). Each sublist PartType# represents one type of particle (e.g. gas, stars, dark matter) and must contain at least the particles coordinates in an N-by-3 matrix \code{Coordinates}. Other optional elements of PartType# are:\cr}

\item{field}{character specifying particle property to concatenate.}

\item{species}{vector of particle species to be included.}
}
\value{
Returns an N-vector or an N-by-3 matrix representing the property "field" of all the requested particles.
}
\description{
Concatenate identical properties of different particle species in a Swift/Gadget snapshot
}
\author{
Danail Obreschkow
}
