% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignHaloTransformation.R
\name{assignHaloTransformation}
\alias{assignHaloTransformation}
\title{Assign transformation metadata to halo properties}
\usage{
assignHaloTransformation(pattern, transformation, exact = FALSE)
}
\arguments{
\item{pattern}{A character string used to match column names in `swift$halos`.}

\item{transformation}{A character string, "position" or "velocity", describing the
transformation-type of the variable.}

\item{exact}{Logical flag. If \code{TRUE}, matches column names exactly;
if \code{FALSE}, uses case-insensitive pattern matching.}
}
\value{
None. Attributes are assigned in-place to columns of `swift$halos`.
}
\description{
Assigns a transformation attribute to one or more columns in `swift$halos`,
based on an exact or partial name match. This is needed to ensure that halo properties are
correctly transformed, e.g. when a halo is unwrapped, centred, or rotated.
}
