% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableSpecies.R
\name{availableSpecies}
\alias{availableSpecies}
\title{Return available particle species}
\usage{
availableSpecies(particle.list = NULL)
}
\arguments{
\item{particle.list}{Optional list of particle data. If `NULL`, it uses `swift$particles`.}
}
\value{
Integer vector of particle species.
}
\description{
This function extracts the particle species from a custom particle list or from
the one stored in `swift$particles` (if no argument is provided). Only entries
with names matching the pattern "PartType#" are considered.
}
