% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{bindHalos}
\alias{bindHalos}
\title{Bind `swift$halos` as local variable `halos`}
\usage{
bindHalos(env = parent.frame())
}
\arguments{
\item{env}{Environment in which to create the binding. Defaults to caller's environment.}
}
\description{
Bind `swift$halos` as local variable `halos`
}
\keyword{internal}
