% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{bindParticles}
\alias{bindParticles}
\title{Bind `swift$particles` as local variable `particles`}
\usage{
bindParticles(env = parent.frame())
}
\arguments{
\item{env}{Environment in which to create the binding. Defaults to caller's environment.}
}
\description{
Bind `swift$particles` as local variable `particles`
}
\keyword{internal}
