% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHaloList.R
\name{checkHaloList}
\alias{checkHaloList}
\title{Check halo list}
\usage{
checkHaloList(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical flag to control whether progress and timing information should be printed in console.}
}
\value{
None. The function stops with an error if any inconsistency is detected.
}
\description{
Verifies the self-consistency of the halo list by checking internal indices and structural properties.
}
\details{
This function inspects the current halo table (`swift$halos`) and throws descriptive errors if inconsistencies are found in the following fields: `NumberOfSubhalos`, `HaloCatalogueIndex`, `HostHaloIndex`, `SubhaloRankByBoundMass`, and `TotalMass`. It assumes that centrals are sorted by decreasing mass and that no orphan satellites are present. No changes are made to the halo table.
}
