% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSwiftData.R
\name{deleteSwiftData}
\alias{deleteSwiftData}
\title{Delete disk image of a `swift` environment}
\usage{
deleteSwiftData(stage = NULL, filename = NULL, verbose = TRUE)
}
\arguments{
\item{stage}{Optional positive integer indicating the processing stage to remove. All higher stages are removed, too. If not given, all processing stages are removed.}

\item{filename}{Optional full path to the temporary file, without extension '.*.img'. If not provided, a simulation-specific default filename,
generated by \code{\link{swiftFileName}}, is adopted.}

\item{verbose}{Logical flag to enable timing and progress messages.}
}
\value{
None.
}
\description{
Deletes the files containing a global `swift` environment previously on the disk,
including associated particle data in subdirectories.
}
\seealso{
\link{saveSwiftData}, \link{loadSwiftData}, \link{swiftFileName}
}
