% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadExamples.R
\name{downloadExamples}
\alias{downloadExamples}
\title{Download and Install the Example Data}
\usage{
downloadExamples(destdir = NULL, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{destdir}{Character string. Destination directory where the data
should be stored. Defaults to the current working directory. If not given, the
path in `swift$.paths$examples` is used.}

\item{overwrite}{Logical. If \code{TRUE}, any existing file of the same
name will be overwritten.}

\item{verbose}{Logical flag to control whether progress and timing information should be printed in console.}
}
\value{
Paths to the downloaded files, invisibly.
}
\description{
This helper downloads the SWIFT example dataset (about 260 MB),
which is too large to be included in the CRAN package bundle.
The data are hosted externally on GitHub, using Git Large File Storage
(see https://git-lfs.com).

By default, the data are downloaded into a user-specified directory,
and the function returns the path to the downloaded files. This is the base
directory that need to be provided when setting the various paths
using \link{setPath}.
}
