% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterHaloList.R
\name{filterHaloList}
\alias{filterHaloList}
\title{Filter halo list}
\usage{
filterHaloList(index, isHaloCatalogueIndex = FALSE, verbose = TRUE)
}
\arguments{
\item{index}{Integer vector of halo indices. By default, these indices are interpreted as row-indices in the halo data.table `swift$halos`.}

\item{isHaloCatalogueIndex}{Logical flag. If \code{TRUE}, the \code{index} values are taken to be the unique halo indices `HaloCatalogueIndex` available in the file `swift$paths$halos` and locally stored in an identically named column of `swift$halos`.}

\item{verbose}{Logical flag to control whether progress and timing information should be printed in console.}
}
\value{
None. Modifies `swift$halos` in place.
}
\description{
Filters the halo list to a specified subset of `index` values, ensuring internal indices remain valid and consistent.
}
\details{
The function updates `swift$halos` by keeping only the selected halos and any required central halos to avoid orphaned satellites. Internal indices (`HostHaloIndex`, `SubhaloRankByBoundMass`, `NumberOfSubhalos`) are recomputed accordingly.
}
