% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVector.R
\name{getVector}
\alias{getVector}
\title{Extract vectors from tables with vector components}
\usage{
getVector(x, name, ignore.case = TRUE, keep.colnames = FALSE)
}
\arguments{
\item{x}{object containing table-like data, such as a data.frames, data.table, matrix, big.matrix, etc.}

\item{name}{A character string used as a pattern to match column names \code{"[name].[number]"}.}

\item{ignore.case}{Logical flag indicating whether the match should ignore case. Defaults to \code{TRUE}.}

\item{keep.colnames}{Logical flag indicating whether to return the result with column names.}
}
\value{
m-by-n matrix, where m is the number of rows in x and n is the number of columns containing the pattern \code{name}.
}
\description{
Reunites vector-components in table columns into a simple matrix
}
\examples{
# simple matrix
m = matrix(seq(15),nrow=5,ncol=3)
colnames(m) = c('a','bb.1','bb.2')
test = list(m=m)

# data.frame
test$df = as.data.frame(m)

# data.table
test$dt = data.table::as.data.table(m)

# big.matrix
test$bm = bigmemory::as.big.matrix(m)

# evaluate all
for (keep.colnames in c(FALSE,TRUE)) {
  for (x in test) {
    print(getVector(x[4:5,],'bb',keep.colnames=keep.colnames))
  }
}

}
