% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialiseSwift.R
\name{initialiseSwift}
\alias{initialiseSwift}
\title{Initialise SWIFT simulation metadata}
\usage{
initialiseSwift(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical flag to print process and timing information in console.}
}
\value{
None. Modifies the global `swift$simulation` object.
}
\description{
Loads simulation metadata from a SWIFT snapshot file, stores it in `swift$simulation`, and determines which particle species are present.
}
\details{
The filename pattern of snapshots must be provided via the list `swift$paths$snapshot`, which can be set using the function \link{setPath}.
}
