% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHaloProperties.R
\name{loadHaloProperties}
\alias{loadHaloProperties}
\title{Load extra halo properties}
\usage{
loadHaloProperties(properties, verbose = TRUE)
}
\arguments{
\item{properties}{A vector of character strings specifying the datasets to be loaded, in the usual HDF5 structure `group/dataset`, `group/group/dataset`, etc.}

\item{verbose}{Logical flag to control whether progress and timing information should be printed in console.}
}
\value{
None. Modifies `swift$halos` in place.
}
\description{
Reads additional halo properties from a HDF5 halo catalogue and appends them to the existing halo table `swift$halos`.
}
\details{
This function updates `swift$halos` by appending new columns specified in `properties`. If the requested properties already exist as columns in `swift$halos`, they are ignored. The new column names are identical to the HDF5 path names in `properties`, except that `/` are replaced by `.`.

The full filename of the halo catalogue must be available in `swift$paths$halos`, which can be set using the function \link{setPath}.
}
