% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSwiftData.R
\name{loadSwiftData}
\alias{loadSwiftData}
\title{Load the `swift` environment from disk}
\usage{
loadSwiftData(filename = NULL, verbose = TRUE)
}
\arguments{
\item{filename}{Optional path+filename to the saved `.img` file. If not provided, a default filename
that uniquely encodes `swift$paths` is used, as in \code{\link{saveSwiftData}}.}

\item{verbose}{Logical flag to enable timing and progress messages.}
}
\value{
Logical value: `TRUE` if the file was found and successfully loaded, `FALSE` otherwise.
}
\description{
Loads the global `swift` environment from a previously saved `.img` file on disk,
restoring all objects and internal state as they were at the time of saving.
This complements the function \code{\link{saveSwiftData}}. If the data on the image file
is identical to that currently in the `swift` environment, it is not reloaded.
}
\seealso{
\link{saveSwiftData}, \link{swiftFileName}
}
