% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortHaloList.R
\name{sortHaloList}
\alias{sortHaloList}
\title{Sort halo list}
\usage{
sortHaloList(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical flag to control whether progress and timing information should be printed in console.}
}
\value{
None. Modifies `swift$halos` in place.
}
\description{
Sorts the subhalo list from most to least massive, ensuring that each central subhalo is directly followed by its associated satellite subhalos.
}
\details{
The input halo table (`swift$halos`) is sorted so that more massive centrals appear earlier, followed by their satellites ordered by subhalo rank (also from most to least massive). The function also updates `HostHaloIndex` and computes a new property `NumberOfSubhalos` for each central.
}
\seealso{
\link{checkHaloList}
}
