% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiftFileName.R
\name{swiftFileName}
\alias{swiftFileName}
\title{Generate default filename for storing swift data}
\usage{
swiftFileName()
}
\value{
A character string with the full path, but no file extension.
}
\description{
Constructs a quasi-unique and reproducible filename for storing the global `swift` environment
using a md5-hash of the file properties (basename, size, timestamp) of the HDF5-files given in `swift$.paths`.
The filename is placed in the directory specified by `swift$.paths$tmp`.

This function is typically used by \code{\link{saveSwiftData}} and \code{\link{loadSwiftData}}
when no filename is explicitly provided.
}
\seealso{
\link{saveSwiftData}, \link{loadSwiftData}
}
