% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformationCols.R
\name{transformationCols}
\alias{transformationCols}
\title{Identify halo columns associated with a specific transformation}
\usage{
transformationCols(transformation)
}
\arguments{
\item{transformation}{A character string specifying the transformation type. Must be
either `"position"` or `"velocity"`.}
}
\value{
A list of three integer vectors corresponding to the column indices for the x, y, and z
components of the selected transformation. If column names are not consistent with this
3D convention, an error is thrown.
}
\description{
Returns the column indices in `swift$halos` corresponding to a given
transformation type, such as `"position"` or `"velocity"`. These columns are assumed to
represent 3D vectors, with components indicated by a trailing `.1`, `.2`, or `.3` in the
column names.
}
