% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHaloProperties.R
\name{addHaloProperties}
\alias{addHaloProperties}
\title{Add new halo properties}
\usage{
addHaloProperties(x, names = NULL, HaloCatalogueIndex = NULL, verbose = FALSE)
}
\arguments{
\item{x}{A vector, matrix, data.frame, or data.table to append as new columns to the data.table `swift$halos`.}

\item{names}{A character string or vector of strings specifying the name(s) of the new column(s). If not given, the column names of \code{x} are used.}

\item{HaloCatalogueIndex}{Optional vector specifying the rows in the halo table where the new properties should be written. If \code{NULL}, all rows are updated, and the order is the same as in \code{x}.}

\item{verbose}{Logical flag to control whether progress and timing information should be printed in the console.}
}
\value{
None. Modifies `swift$halos` in place.
}
\description{
Adds new halo properties to the halo table stored in `swift$halos`.
}
