% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSwiftData.R
\name{deleteSwiftData}
\alias{deleteSwiftData}
\title{Delete disk image of a `swift` environment}
\usage{
deleteSwiftData(filename = NULL)
}
\arguments{
\item{filename}{Optional path+filename to the saved `.img` file. If not provided, a default filename
that uniquely encodes `swift$paths` is used, as in \code{\link{saveSwiftData}}.}
}
\value{
None.
}
\description{
Deletes the files containing a global `swift` environment previously on the disk,
including associated particle data in subdirectories.
}
\seealso{
\link{saveSwiftData}, \link{loadSwiftData}, \link{swiftFileName}
}
