% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCell.R
\name{fetchCell}
\alias{fetchCell}
\title{Fetch particles associated with one or more cells}
\usage{
fetchCell(
  index,
  unwrap = TRUE,
  properties = NULL,
  species = NULL,
  verbose = FALSE
)
}
\arguments{
\item{index}{Integer vector of cell indices, interpreted as row-index in the cell table `swift$cells`.}

\item{unwrap}{Logical flag to control whether the objects are unwrapped using periodic boundary conditions. Requires size of simulation box to be specified in `swift$simulation$BoxSize`. If multiple cells are loaded via a vector index, they are all unwrapped collectively, not cell-by-cell.}

\item{properties}{Vector of character strings specifying the properties to be extracted. If `NULL` all available properties are returned. For vector properties, provide the vector name (e.g. `Coordinates`) without component index (e.g. `Coordinates.1`).}

\item{species}{Integer vector of particle species to include. Recognized values are:
`0` (gas), `1` (dark matter), `4` (stars), and `5` (black holes). If `NULL`, all available species are used.}

\item{verbose}{Logical flag to enable timing and progress messages.}
}
\value{
A list of particle matrices, one for each selected species (e.g., `PartType1`, `PartType4`, etc.).
}
\description{
Retrieves the particles associated with one or more cells from the currently loaded data.
}
\details{
This function assumes that cell and particle data are already available in
`swift$cells` and `swift$particles$cells` objects, respectively.
}
