% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchHalo.R
\name{fetchHalo}
\alias{fetchHalo}
\title{Fetch particles associated with one or more halos}
\usage{
fetchHalo(
  index,
  isHaloCatalogueIndex = FALSE,
  unwrap = TRUE,
  properties = NULL,
  substructure = FALSE,
  species = NULL,
  verbose = FALSE
)
}
\arguments{
\item{index}{Integer halo index. By default, this index is interpreted as row-index in the halo table `swift$halos`.}

\item{isHaloCatalogueIndex}{Logical flag. If \code{TRUE}, the \code{index} value is taken to be the unique `HaloCatalogueIndex` available in the file `swift$.paths$halos` and locally stored in an identically named column of `swift$halos`.}

\item{unwrap}{Logical flag to control whether the objects are unwrapped using periodic boundary conditions. Requires size of simulation box to be specified in `swift$simulation$BoxSize`.}

\item{properties}{Vector of character strings specifying the properties to be extracted. If `NULL` all available properties are returned. For vector properties, provide the vector name (e.g. `Coordinates`) without component index (e.g. `Coordinates.1`).}

\item{substructure}{Logical flag for including substructure. If \code{TRUE} and if the requested index refers to a central subhalo, partlcies from its satellite subhalos are included.}

\item{species}{Integer vector of particle species to include. Recognized values are:
`0` (gas), `1` (dark matter), `4` (stars), and `5` (black holes). If `NULL`, all available species are used.}

\item{verbose}{Logical flag to enable timing and progress messages.}
}
\value{
A list of particle matrices, one for each selected species (e.g., `PartType1`, `PartType4`, etc.).
The list is assigned the class `snapshot` for compatibility with downstream functions, such as `plot`.
}
\description{
Retrieves the particles associated with one or more halos from the currently loaded data.
}
\details{
This function assumes that halo and particle data have already been loaded into memory
via the `swift$halos` and `swift$particles` objects, respectively.
}
