% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHaloList.R
\name{loadHaloList}
\alias{loadHaloList}
\title{Load basic halo list}
\usage{
loadHaloList(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical flag to control whether progress and timing information should be printed in console.}
}
\value{
None. The loaded halo list is stored in `swift$halos`.
}
\description{
Loads basic halo properties from an HDF5 file and constructs a `data.table` with key columns required for further analysis.
}
\details{
Creates a `data.table` of halo properties accessible via `swift$halos`.
Each row represents a subhalo, which is either a central subhalo (excluding substructure)
or a satellite subhalo (associated with a central). The table includes the following columns:

The full filename of the halo catalogue must be available in `swift$paths$halos`, which can be set using the function \link{setPath}.


\itemize{
  \item \code{HaloCatalogueIndex}: Unique subhalo index.
  \item \code{TotalMass}: Total mass of all particles in the subhalo, in the units of the halo file.
  \item \code{HostHaloIndex}: For satellite subhalos, this is the row-index of the associated central subhalo.
        For central subhalos, this is 0.
  \item \code{SubhaloRankByBoundMass}: For satellite, this is the rank of the subhalo among the satellites
        belonging to the same central, from most to least massive. For centrals, this is 0.
}
}
