% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHaloList.R
\name{loadHaloList}
\alias{loadHaloList}
\title{Load basic halo list}
\usage{
loadHaloList(filename, verbose = TRUE)
}
\arguments{
\item{filename}{Full path of the HDF5 halo file.}

\item{verbose}{Logical flag to control whether progress and timing information should be printed in console.}
}
\value{
None. The loaded halo list is stored in `swift$halos`.
}
\description{
Loads basic halo properties from an HDF5 file and constructs a `data.table` with key columns required for further analysis.
}
\details{
Creates a `data.table` of halo properties and assigns it to `swift$halos` via an active binding for convenient in-scope access using the variable `halos`.
}
