% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadParticles.R
\name{loadParticles}
\alias{loadParticles}
\title{Load and regroup particles from snapshot files}
\usage{
loadParticles(
  method,
  species = NULL,
  properties = c("Masses", "Coordinates", "Velocities"),
  verbose = TRUE
)
}
\arguments{
\item{method}{One of the following character strings
\itemize{
  \item \code{"halos"}: Loads all the particles associated with the subhalos stored in the data.frame `swift$halos`, normally created earlier via \link{loadHaloList}. Subhalo associations are red from membership files, whose path must be set in `swift$.paths$membership`.
  \item \code{"cells"}: Loads all the particles specified in the data.frame `swift$cells`, which must contain columns named `x`, `y`, `z`, specifying the cell centres, as well as columns named `width` and/or `radius` specifying, either the side-length of cubic cells or the radius of spherical ones. Such a data.frame can be generated using \link{makeGridCells}.
}}

\item{species}{Integer vector of particle types to include. Recognized values are: `0` (gas), `1` (dark matter), `2` (background), `3` sink, `4` (stars), `5` (black holes), and `6` (neutrinos). Defaults to all four.}

\item{properties}{Character vector of particle properties to load. Defaults to `c("Masses", "Coordinates", "Velocities")`. These properties have to exist in the snapshot files, except for the property `Rank_bound`, which is assumed to exist in the membership file.}

\item{verbose}{Logical flag to enable timing and progress messages.}
}
\value{
None. The loaded particle data is stored in the `swift$particles` list. Pointers to the particle data are added as new columns to `swift$halos` or `swift$cells` (see details).
}
\description{
This function loads properties of particles of a SWIFT simulation snapshot, accounting for multiple subvolumes if present. The particles are filtered and regrouped according to the \code{method} argument.
}
\details{
Particle data is read from multiple subvolume files.\cr

Particle data are written into the list `swift$particles`, which contains a separate structure `PartType#` for each particle species.
Individual particle properties of a given species are written into a big matrix `swift$particles$PartType#$data`,
created via the \code{bigmemory} package. These matrices (one per species) are stored on the disk in the directory
`swift$.paths$tmp/"particledata"` with filenames specified in `swift$particles$PartType#$filename`. Data are
automatically loaded into the RAM when needed in a fast and memory-lean manner.\cr

Each particle occupies one row in `swift$particles$PartType#$data`, and successive groups of particles correspond to successive rows in `swift$halos` or `swift$cells`.\cr

For each species `#`, two new columns are added to `swift$halos`: `npart.#` is the number of bound particles of that species in each subhalo, and `ipart.#` gives the 1-based index of the first corresponding particle
in the matrix `swift$particles$PartType#$data`.
These pointers allow efficient access to all particles belonging to a given subhalo or cell. In the case of subhalos, if \link{sortHaloList} is called
before loading the particles, satellite subhalos immediately follow their centrals.
In this case, the particle data of each halo, including its substructure, appear contiguously in
the particle matrices.\cr

All required file paths and filename patterns must be provided via the list `swift$.paths`, using the fields `snapshot`, `membership`, and `tmp`. These paths can be set using the function \link{setPath}.
}
\seealso{
\link{setPath}, \link{sortHaloList}
}
