% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSwiftData.R
\name{loadSwiftData}
\alias{loadSwiftData}
\title{Load the `swift` environment from disk}
\usage{
loadSwiftData(stage = NULL, filename = NULL, verbose = TRUE)
}
\arguments{
\item{stage}{Optional positive integer indicating the processing stage. If not provided, the highest existing processing stage is loaded.}

\item{filename}{Optional full path to the temporary file, without extension '.*.img'. If not provided, a simulation-specific default filename,
generated by \code{\link{swiftFileName}}, is adopted.}

\item{verbose}{Logical flag to enable timing and progress messages.}
}
\value{
None. The stage of the loaded data is stored in `swift$stage`
}
\description{
Loads the global `swift` environment from a previously saved `.img` file on disk,
restoring all objects and internal state as they were at the time of saving.
This complements the function \code{\link{saveSwiftData}}. If the data on the image file
is identical to that currently in the `swift` environment, it is not reloaded.
}
\seealso{
\link{saveSwiftData}, \link{deleteSwiftData}, \link{swiftFileName}
}
