% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeHaloProperties.R
\name{removeHaloProperties}
\alias{removeHaloProperties}
\title{Remove columns from the halo table}
\usage{
removeHaloProperties(names, verbose = FALSE)
}
\arguments{
\item{names}{Character vector of column names to remove.}

\item{verbose}{Logical flag to control whether progress and timing information should be printed in the console.}
}
\value{
None. Modifies `swift$halos` in place.
}
\description{
Removes one or more named columns from the `swift$halos` table.
}
