% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSwiftData.R
\name{saveSwiftData}
\alias{saveSwiftData}
\title{Save the `swift` environment to disk}
\usage{
saveSwiftData(stage = NULL, filename = NULL, verbose = TRUE, force = FALSE)
}
\arguments{
\item{stage}{Optional positive integer indicating the new processing stage. If not provided, this is set to `swift$stage+1`.}

\item{filename}{Optional full path to the temporary file, without extension '.*.img'. If not provided, a simulation-specific default filename,
generated by \code{\link{swiftFileName}}, is adopted.}

\item{verbose}{Logical flag to enable timing and progress messages.}

\item{force}{Logical flag. If TRUE, an R-image is saved even if the data hasn't changed since the last backup.}
}
\value{
None. Writes a `.img` file to disk containing the full `swift` environment, with pointers to linked particle data if available, as well as a `.hash` file containing a unique identifier of the img-file.
}
\description{
Saves the global `swift` environment, except for objects with names beginning with '.',
such as the default object .paths, which is deliberately excluded to allow for
paths to change without requiring to recompute the data. The data are saved
in a file `[filename].img` using R's `save()` function, allowing reloading the entire
simulation context in a later session. An ascii file `[filename].hash` is also
produced which contains a hash for the `swift` environment to later assess
if the currently loaded data is identical to the saved one.
}
\seealso{
\link{loadSwiftData}, \link{deleteSwiftData}, \link{swiftFileName}
}
