% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPath.R
\name{setPath}
\alias{setPath}
\title{Set a named file path}
\usage{
setPath(name, path)
}
\arguments{
\item{name}{A character string specifying the name of the path to set. Examples include:
\itemize{
  \item \code{"halos"}: path to the HDF5 halo file, including `.hdf5` extension
  \item \code{"particles"}: path to the HDF5 particle file, including `.hdf5` extension, but without subvolume numbers (`.0`, `.1`, ...).
  \item \code{"tmp"}: path to the directory containing temporary R-img files for fast access
}}

\item{path}{A character string specifying the full file or directory path.}
}
\value{
None. The named path is stored in `swift$paths[[name]]`.
}
\description{
Stores a file path in `swift$paths` under a specified name.
These paths are used by other functions that accept `filename = NULL`.
}
