% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSwift.R
\name{setSwift}
\alias{setSwift}
\title{Set a named value in a swift field}
\usage{
setSwift(name, value)
}
\arguments{
\item{name}{A character string specifying the name/key to be written. Sub-lists can be specified using the `$` symbol.}

\item{value}{The value to store under `swift[[name]]`. Can be a path, object, or structure depending on context.}
}
\value{
None. The value is stored in `swift[[name]]`.
}
\description{
Stores a value in the global `swift` object under a specified name.
This function is used primarily to write values into the protected (read-only)
fields of swift, such as `swift$paths`, `swift$halos`, `swift$particles`, `swfit$simulation`.
}
\examples{
setSwift("simulation$BoxSize", 100)

}
\seealso{
\link{setPath}
}
