% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSwift.R
\name{setSwift}
\alias{setSwift}
\title{Set a named value in a swift field}
\usage{
setSwift(name, value)
}
\arguments{
\item{name}{A character string specifying the name/key to be written. Sub-lists can be specified using the `$` symbol.}

\item{value}{The value to store under `swift[[name]]`. Can be a path, object, or structure depending on context.}
}
\value{
None. The value is stored in `swift[[name]]`.
}
\description{
Stores a value in the global `swift` object under a specified name.
This function is used to write values into the protected (read-only) fields of swift,
such as `swift$backup`, `swift$cells`, `swift$halos`, `swift$particles`, `swfit$simulation`, `swift$.paths`.
}
\examples{
setSwift("simulation$BoxSize", 100)
# but use normal access to write into non-protexted fields
swift$newField <- list(text='abc', value=123)
swift$newField$extra <- diag(3)

}
\seealso{
\link{setPath}
}
