\name{SWIFT-package}
\alias{SWIFT-package}
\alias{SWIFT}
\docType{package}
\title{
   \packageTitle{SWIFT}
}
\description{
  \packageDescription{SWIFT}
}
\details{	
  The DESCRIPTION file:
  \packageDESCRIPTION{SWIFT}
  \packageIndices{SWIFT}
The SWIFT model contains multipe functions that will help the user to track and 
understand stable water isotopic fluctuations within the transport tissue of a 
plant. The central functions SWIFT_SB and SWIFT_H provide isotopic signatures at
(i) the stem base throughout the entire study time and (ii) at a height and 
timing definined by the user, respectively. 
Additional functions are SoilRootCond, PSI0calc and Bprep, which in this order 
help the user to define (i) the soil to root conductivity for every defined soil
layer, (ii) the water potential at stem base for every timestep, and (iii) the 
root length distribution for every defined soil layer.
}
\author{	
  \packageAuthor{SWIFT}
  Maintainer: 
  \packageMaintainer{SWIFT}
}
\references{
De Deurwaerder et al (IN REVIEW)
}