% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift.R
\docType{data}
\name{swift}
\alias{swift}
\title{Handle Astrophysical Simulation Data from SWIFT/SOAP}
\description{
`swift` is both the name of this package and of its main user-facing environment.

This package bundles routines for accessing and visualising
astrophysical simulation data output by the hydro-gravitational
N-body simulation code SWIFT (SPH With Inter-dependent Fine-grained Tasking),
as well as value-added data generated with SOAP (Spherical Overdensity Aperture Processor). The current primary use is the post-processing of
the Colibre and Flamingo cosmological simulations for research purposes.
Within some limits, the routines can also be used to access
and render data generated by Gadget (GAlaxies with Dark matter and Gas intEracT).
A mini-tutorial is provided in the example section below.

Data are store locally in the `swift` package and environment.
}
\examples{
\donttest{# The following example illustrates most of the package’s main routines.

# Caution: This example includes downloading data from github and
# slightly time-consuming post-processing steps. All in all, this
# may take a or longer minute to execute on a slower system. To customize paths and
# obtain intermediate progress updates, it may be convenient to copy this code
# into a new R-script and run it from there.

# plotting options
show.full.box = TRUE
show.halo = TRUE

# set global paths (customize these as desired)
setPath('examples',path.expand("~/Data/SwiftExamples/"))
setPath('tmp',path.expand("~/Data/SwiftTmp/"))

# optionally download data
downloadExamples()

# set paths to specific simulation snapshot
base = paste0(swift$.paths$examples,'colibre/L0012N0094/')
setPath('halos',paste0(base,'halo_properties_0127.hdf5'))
setPath('snapshot',paste0(base,'colibre_0127.\%d.hdf5'))
setPath('membership',paste0(base,'membership_0127.\%d.hdf5'))

# load latest local post-processed data for the paths given above
loadSwiftData()

# load initialize data
if (isSwiftStage(0)) {
  clearSwift() # delete existing data in swift environment, if any
  initialiseSwift() # initialise simulation
  saveSwiftData() # save environment
}

# partition simulation box and load all particles
if (isSwiftStage(1)) {
  makeGridCells() # partition the simulation in a cartesian grid of cubic cells
  loadParticles('cells') # load all particles
  saveSwiftData() # save environment
}

# plot full simulation box
if (show.full.box) {
  x = fetchCell(1:512, unwrap=F,verbose=T)
  par(mar=rep(0.1,4))
  render(x, npixels=400,center=6.25,width=12.5,gamma=.7)
}

# make custom halo list
if (isSwiftStage(2)) {
  loadHaloList() # load master list of halos
  filterHaloList(swift$halos$TotalMass>1e1) # apply basic selection function
  sortHaloList() # sort central subhalos by decreasing mass, satellites follow their centrals
  checkHaloList() # check for self consistency
  saveSwiftData() # save environment
}

# load particles in selected halos
if (isSwiftStage(3)) {
  loadParticles('halos', species=c(0,4)) # loads all gas and star particles
  saveSwiftData() # save environment
}

# assign additional halo properties
if (isSwiftStage(4)) {
  loadHaloProperties(c('InputHalos/HaloCentre','ExclusiveSphere/50kpc/StellarMass'))
  addHaloProperties(seq(nrow(swift$halos)), 'my.property') # add custom halo properties
  assignHaloTransformation('centre','position') # assign transformation properties to halos
  saveSwiftData() # save environment
}

# plot one halo
if (show.halo) {
  x = fetchHalo(1, species=c(0,4), substructure=TRUE, verbose=TRUE)
  par(mar=rep(0.1,4))
  render(x, npixels=400, width=1, center=getVector(x$halos[1,],'InputHalos.HaloCentre'))
}

# post-processed data can be deleted using
# > deleteSwiftData()
# or all data up from a certain stage onward using
# > deleteSwiftData([stage index])
}

}
\seealso{
More examples of rending galaxies are provided in the documentation of \link{render}.
}
\author{
Danail Obreschkow
}
\keyword{datasets}
