% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift.env.R
\docType{data}
\name{swift}
\alias{swift}
\title{The `swift` environment}
\format{
An object of class \code{environment} of length 4.
}
\usage{
swift
}
\description{
An environment that stores loaded simulation data, such as `halos`, `particles`, and file `paths`.
The entries in this environment are exposed to the user as read-only bindings.
Internally, values are stored in a hidden environment and modified via package functions.
}
\keyword{datasets}
