% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift.R
\docType{data}
\name{swift}
\alias{swift}
\title{Basic handling of SWIFT/SOAP simulation data}
\format{
An object of class \code{environment} of length 5.
}
\usage{
swift
}
\description{
`swift` is both the name of this package and of its main user-facing environment.

The swift package Routines to handle astrophysical simulation data output by the hydro-gravitational N-body simulation code SWIFT (SPH With Inter-dependent Fine-grained Tasking), as well as value-added data generated with SOAP (Spherical Overdensity Aperture Processor).

Data are store locally in the `swift` package and environment.
}
\examples{
\donttest{
# The following example uses most of the package’s main routines.
#
# set path where temporary fast-access files are stored
setPath('tmp','/Users/00077389/Data/tmp/')

# set paths to simulation data
setPath('halos',system.file("colibre", "halo_properties_0127.hdf5", package = "swift"))
setPath('snapshot',gsub('\\\\.0\\\\.','.\%d.',
        system.file("colibre", "colibre_0127.0.hdf5", package = "swift")))
setPath('membership',gsub('\\\\.0\\\\.','.\%d.',
        system.file("colibre", "membership_0127.0.hdf5", package = "swift")))

# load post-processed simulation data, if available
loadSwiftData()

# initialise post-processing and load/filter halo list with linking properties
if (isSwiftStage(0)) {

  # delete existing data in swift environment, if any
  clearSwift()

  # initialise simulation
  initialiseSwift()

  # load basic halo list
  loadHaloList()

  # apply basic selection function
  filterHaloList(swift$halos$TotalMass>10)

  # sort central subhalos from most massive to least massive with satellites
  # following their centrals, ordered by their own ranked mass
  sortHaloList()

  # check for self consistency
  checkHaloList()

  # save halos
  saveSwiftData()

}

# load particle data of selected halos
if (isSwiftStage(1)) {

  # load particles data
  loadParticlesInHalos()

  # save swift environment for later quick assess via loadSwiftData()
  saveSwiftData()

}

# load/add additional halo properties and assign transformation rules
if (isSwiftStage(2)) {

  # populate halo list with additional properties from file
  properties = 'InputHalos/HaloCentre'
  for (g1 in c('ExclusiveSphere','InclusiveSphere')) {
    for (g2 in c('3kpc','10kpc','30kpc','50kpc','100kpc')) {
      for (g3 in c('TotalMass','StellarMass','AtomicHydrogenMass','CentreOfMass')) {
        properties = c(properties,paste(g1,g2,g3,sep='/'))
      }
    }
  }
  loadHaloProperties(properties)
  rm(properties,g1,g2,g3)

  # add custom halo properties
  addHaloProperties(seq(nrow(swift$halos)), 'my.property')

  # assign transformation properties to halos
  # this will ensure that these halo properties are also transformed,
  # if the particles are transformed in a certain way, e.g. un-wrapped around
  # periodic boundaries
  assignHaloTransformation('centre','position')

  # save swift environment for later quick assess via loadSwiftData()
  saveSwiftData()

}

# extract particle data of single halo
x = fetchHalo(1, species=c(0,4), substructure=TRUE, verbose=TRUE)

# render halo
par(mar=rep(0.1,4))
library(gadgetry)
plot(x, npixels=100, width=1, center=getVector(x$Halos[1,],'InputHalos.HaloCentre'))

# delete all data using
# deleteSwiftData()
# or all data up from a certain stage onward using
# deleteSwiftData(...)
}

}
\keyword{datasets}
