% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift.env.R
\docType{data}
\name{swift}
\alias{swift}
\title{Package environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
swift
}
\description{
Environment used to store currently loaded simulation data.
}
\author{
Danail Obreschkow
}
\keyword{datasets}
